pyproject.toml
[project]
name = "alloy-replication"
version = "0.1.0"
description = "Replication of 'On the Effectiveness of LLMs in Writing Alloy Formulas' (Hong et al., 2025)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
